dashboard/requirements.txt
# Trading Dashboard V4.2 Dependencies

# Flask web framework
flask>=2.3.0
flask-cors>=4.0.0

# Environment variables
python-dotenv>=1.0.0

# Broker APIs
alpaca-py>=0.13.0
oandapyV20>=0.6.3

# Cryptography for Kalshi auth
cryptography>=46.0.5

# HTTP requests
requests>=2.31.0
aiohttp>=3.9.0

# Data handling
pandas>=2.0.0
